<FIX> count query twice

+++ src/main/java/com/instream/tenant/domain/common/domain/dto/PaginationInfoDto.java
@@ -11,7 +11,7 @@ import lombok.experimental.SuperBuilder;
 public class PaginationInfoDto<T extends CollectionDto> extends PaginationDto<T> {
     @Schema(description = "총 페이지 개수, 0부터 시작")
     @Min(value = 0)
-    int pageCount;
+    long pageCount;
 
     @Schema(description = "총 개수, 0부터 시작")
     @Min(value = 0)
